object-cache: move validate_key into the library crate with tests
Move key/prefix validation out of the object-cache-srv binary (which has
no test target) into micromegas-object-cache::validation, and put its
unit tests under that crate's tests/ folder. HTTP-specific helpers
(parse_range_header, is_not_found) stay in the srv.
