+++ rust/object-cache/src/lib.rs
@@ -3,6 +3,7 @@ pub mod blocks;
 pub mod client;
 pub mod memory_backend;
 pub mod range_cache;
+pub mod validation;
 
 #[cfg(feature = "foyer")]
 pub mod foyer_backend;
